backend/pyproject.toml
[project]
name = "vue-fastapi-admin"
version = "0.1.0"
description = "Vue Fastapi admin"
authors = [
    {name = "mizhexiaoxiao", email = "mizhexiaoxiao@gmail.com"},
]
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.20,<1",
    "email-validator>=2.2,<3",
    "fastapi>=0.111,<1",
    "httpx>=0.28,<1",
    "loguru>=0.7,<1",
    "passlib>=1.7,<2",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "pyjwt>=2.10,<3",
    "python-dotenv>=1.0,<2",
    "python-multipart>=0.0.20,<1",
    "tortoise-orm>=0.23,<1",
    "uvicorn>=0.34,<1",
    "websockets>=15,<16",
]

[tool.black]
line-length = 120
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 120
lint.extend-select = []
lint.ignore = [
    "F403",
    "F405",
]

[tool.aerich]
tortoise_orm = "app.settings.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
